w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/drawing-images-to-the-canvas/2d.drawImage.3arg-expected.txt

2d.drawImage.3arg
Actual output:
Expected output:

 

FAIL Canvas test: 2d.drawImage.3arg assert_approx_equals: Green channel of the pixel at (0, 0) expected 255 +/- 2 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:48:25
http://localhost:8800/2dcontext/drawing-images-to-the-canvas/2d.drawImage.3arg.html:28:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
